requirement.txt
llama-index-core==0.12.*
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
llama-index-postprocessor-flashrank
chromadb
markitdown
streamlit